+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 - fixed memory violations found by the new added ASan-memory-check
 - some cases of invalid http-requests had returned an 200 reponse with empty body, which was fixed
 - changed token-key in helm-chart from configmap to secret
+- fixed memory-corruption in cluster-resize, because a pointer was used after free
 
 ### Removed